part_4/rearc_cdk_project/requirements.txt
attrs==25.3.0
aws-cdk-lib==2.184.1
aws-cdk.asset-awscli-v1==2.2.227
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==40.7.0
beautifulsoup4==4.13.3
boto3==1.37.12
botocore==1.37.12
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
constructs==10.4.2
idna==3.10
importlib_resources==6.5.2
jmespath==1.0.1
jsii==1.109.0
numpy==2.2.3
pandas==2.2.3
publication==0.0.3
py4j==0.10.9.7
pyspark==3.5.5
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
soupsieve==2.6
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
